+++ packages/agtree/test/parser/cosmetic/scriptlet-body/abp-scriptlet.test.ts
@@ -7,12 +7,14 @@ import {
 
 import { NodeExpectContext, type NodeExpectFn } from '../../../helpers/node-utils';
 import { type ScriptletInjectionRuleBody } from '../../../../src/nodes';
-import { AbpSnippetInjectionBodyParser } from '../../../../src/parser/cosmetic/body/abp-snippet-injection-body-parser';
+import {
+    AbpSnippetInjectionBodyParser,
+} from '../../../../src/parser/cosmetic/scriptlet-body/abp-snippet-injection-body-parser';
 import { AdblockSyntaxError } from '../../../../src/errors/adblock-syntax-error';
 import { EMPTY, SPACE } from '../../../../src/utils/constants';
 import {
     AbpSnippetInjectionBodyGenerator,
-} from '../../../../src/generator/cosmetic/body/abp-snippet-injection-body-generator';
+} from '../../../../src/generator/cosmetic/scriptlet-body/abp-snippet-injection-body-generator';
 import { AbpSnippetInjectionBodyCommon } from '../../../../src/common/abp-snippet-injection-body-common';
 
 describe('AbpSnippetInjectionBodyParser', () => {