+++ packages/agtree/test/parser/cosmetic/scriptlet-body/ubo-scriptlet.test.ts
@@ -9,11 +9,11 @@ import { NodeExpectContext, type NodeExpectFn } from '../../../helpers/node-util
 import { type ScriptletInjectionRuleBody } from '../../../../src/nodes';
 import {
     UboScriptletInjectionBodyParser,
-} from '../../../../src/parser/cosmetic/body/ubo-scriptlet-injection-body-parser';
+} from '../../../../src/parser/cosmetic/scriptlet-body/ubo-scriptlet-injection-body-parser';
 import { AdblockSyntaxError } from '../../../../src/errors/adblock-syntax-error';
 import {
     UboScriptletInjectionBodyGenerator,
-} from '../../../../src/generator/cosmetic/body/ubo-scriptlet-injection-body-generator';
+} from '../../../../src/generator/cosmetic/scriptlet-body/ubo-scriptlet-injection-body-generator';
 
 describe('UboScriptletInjectionBodyParser', () => {
     describe('UboScriptletInjectionBodyParser.parse - valid cases', () => {